back-end/requirements.txt
Flask==2.3.2
flask-jwt-extended==4.4.2
flask-sqlalchemy==3.0.3
flask-cors==3.0.10
python-dotenv==1.0.0
pyjwt==2.6.0
bcrypt==4.0.1